+++ common/src/main/java/ru/yjailbir/commonslib/dto/AccountDto.java
@@ -1,4 +1,4 @@
 package ru.yjailbir.commonslib.dto;
 
-public record AccountDto(String currency, String name, Integer balance, boolean active) {
+public record AccountDto(String currency, String name, Double balance, boolean active) {
 }
